+++ modded/v1_20_6/fabric/src/main/java/org/geysermc/geyser/adapters/fabric/v1_20_6/FabricCommandManagerAdapter_v1_20_6.java
@@ -23,42 +23,21 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-plugins {
-    id("adapters.modded-conventions")
-}
+package org.geysermc.geyser.adapters.fabric.v1_20_6;
 
-val neoVersion = "21.5.91"
-val metaProperties = minecraftVersion("1.21.5", mapOf("neo_version" to neoVersion))
+import net.minecraft.commands.CommandSourceStack;
+import org.geysermc.geyser.adapters.modded.v1_20_6.CommandManagerAdapter_v1_20_6;
+import org.incendo.cloud.CommandManager;
+import org.incendo.cloud.SenderMapper;
+import org.incendo.cloud.execution.ExecutionCoordinator;
+import org.incendo.cloud.fabric.FabricServerCommandManager;
 
-java {
-    sourceCompatibility = JavaVersion.VERSION_21
-    targetCompatibility = JavaVersion.VERSION_21
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-architectury {
-    platformSetupLoomIde()
-    neoForge()
-}
-
-dependencies {
-    neoForge("net.neoforged:neoforge:$neoVersion")
-
-    api(project(":modded_common_v1_21_5", configuration = "namedElements"))
-    shadowCommon(project(":modded_common_v1_21_5", configuration = "transformProductionNeoForge"))
-
-    compileOnly(project(":common"))
-    modImplementation("org.incendo:cloud-neoforge:2.0.0-beta.12")
-    include("org.incendo:cloud-neoforge:2.0.0-beta.12")
-}
-
-tasks {
-    withType<ProcessResources> {
-        filteringCharset = "UTF-8"
-        filesMatching("META-INF/neoforge.mods.toml") {
-            expand(metaProperties)
-        }
+public class FabricCommandManagerAdapter_v1_20_6 extends CommandManagerAdapter_v1_20_6 {
+    @Override
+    public <T> CommandManager<T> getCommandManager(SenderMapper<CommandSourceStack, T> mapper) {
+        return new FabricServerCommandManager<>(
+                ExecutionCoordinator.simpleCoordinator(),
+                mapper
+        );
     }
 }
