+++ modded/v1_20_6/fabric/src/main/java/org/geysermc/geyser/adapters/fabric/v1_20_6/Fabric_v1_20_6.java
@@ -23,42 +23,17 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-plugins {
-    id("adapters.modded-conventions")
-}
+package org.geysermc.geyser.adapters.fabric.v1_20_6;
 
-val neoVersion = "21.5.91"
-val metaProperties = minecraftVersion("1.21.5", mapOf("neo_version" to neoVersion))
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import org.geysermc.geyser.adapters.CommandManagerAdapter;
+import org.geysermc.geyser.adapters.modded.v1_20_6.GeyserAdapter_v1_20_6;
 
-java {
-    sourceCompatibility = JavaVersion.VERSION_21
-    targetCompatibility = JavaVersion.VERSION_21
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-architectury {
-    platformSetupLoomIde()
-    neoForge()
-}
-
-dependencies {
-    neoForge("net.neoforged:neoforge:$neoVersion")
-
-    api(project(":modded_common_v1_21_5", configuration = "namedElements"))
-    shadowCommon(project(":modded_common_v1_21_5", configuration = "transformProductionNeoForge"))
-
-    compileOnly(project(":common"))
-    modImplementation("org.incendo:cloud-neoforge:2.0.0-beta.12")
-    include("org.incendo:cloud-neoforge:2.0.0-beta.12")
-}
-
-tasks {
-    withType<ProcessResources> {
-        filteringCharset = "UTF-8"
-        filesMatching("META-INF/neoforge.mods.toml") {
-            expand(metaProperties)
-        }
+public class Fabric_v1_20_6 implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        ServerLifecycleEvents.SERVER_STARTING.register(GeyserAdapter_v1_20_6::setServer);
+        CommandManagerAdapter.set(new FabricCommandManagerAdapter_v1_20_6());
     }
 }
