+++ modded/v1_20_6/fabric/src/main/resources/fabric.mod.json
@@ -14,13 +14,13 @@
     "environment": "*",
     "entrypoints": {
         "main": [
-            "org.geysermc.geyser.adapters.fabric.v1_21_5.Fabric_v1_21_5"
+            "org.geysermc.geyser.adapters.fabric.v1_20_6.Fabric_v1_20_6"
         ]
     },
     "depends": {
         "fabricloader": ">=0.16.7",
         "fabric-api": "*",
-        "minecraft": ">=${minecraft_version}",
+        "minecraft": ">=${minecraft_version} <=1.21.4",
         "geyser-fabric": "*"
     }
 }
