+++ modded/v1_20_6/neoforge/gradle.properties
