--- modded/v1_21_1/common/src/main/java/org/geysermc/geyser/adapters/modded/v1_21_1/GeyserAdapter_v1_21_1.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2019-2025 GeyserMC. http://geysermc.org
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @author GeyserMC
- * @link https://github.com/GeyserMC/Geyser
- */
-
-package org.geysermc.geyser.adapters.modded.v1_21_1;
-
-import net.minecraft.server.MinecraftServer;
-
-public class GeyserAdapter_v1_21_1 {
-    private static MinecraftServer server;
-
-    public static void setServer(MinecraftServer server) {
-        if (GeyserAdapter_v1_21_1.server != null) throw new IllegalStateException("Server is already defined.");
-
-        GeyserAdapter_v1_21_1.server = server;
-    }
-
-    public static MinecraftServer getServer() {
-        return server;
-    }
-}
