--- modded/v1_21_1/fabric/src/main/resources/fabric.mod.json
@@ -1,25 +1,0 @@
-{
-    "schemaVersion": 1,
-    "id": "${id}",
-    "version": "${version}",
-    "name": "${name}",
-    "description": "${description}",
-    "authors": ["${author}"],
-    "contact": {
-        "website": "${website}",
-        "repo": "https://github.com/GeyserMC/geyser-adapters"
-    },
-    "license": "${license}",
-    "icon": "assets/geyser/icon.png",
-    "environment": "*",
-    "entrypoints": {
-        "main": [
-            "org.geysermc.geyser.adapters.fabric.v1_21_1.Fabric_v1_21_1"
-        ]
-    },
-    "depends": {
-        "fabricloader": ">=0.16.7",
-        "fabric-api": "*",
-        "minecraft": ">=${minecraft_version}"
-    }
-}
