--- modded/v1_21_1/neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,31 +1,0 @@
-modLoader="javafml"
-loaderVersion="[1,)"
-license="${license}"
-
-[[mods]]
-modId="${id}"
-version="${version}"
-displayName="${name}"
-displayURL="${website}"
-authors="${author}"
-description='''${description}'''
-
-#[[mixins]]
-#config="${id}.mixins.json"
-
-#[[accessTransformers]]
-#file="META-INF/accesstransformer.cfg"
-
-[[dependencies.${id}]]
-modId="neoforge"
-type="required"
-versionRange="[${neo_version},)"
-ordering="NONE"
-side="BOTH"
-
-[[dependencies.${id}]]
-modId="minecraft"
-type="required"
-versionRange="[${minecraft_version},)"
-ordering="NONE"
-side="BOTH"
